move assert_error_matches to test_util
